BJS2-66382 1. Объединение редис конфигурации 2. Удаление ненужной конфигурации
