--- lombok.config
@@ -1,1 +1,0 @@
-lombok.copyableAnnotations += org.springframework.beans.factory.annotation.Qualifier
